Renames performance module to `accumulation`
The new name `accumulation` more accurately reflects the module's responsibility for gathering and managing values, especially within the context of robust numerical handling and preventing overflow.
