+++ packages/ec-core/src/performance/accumulation.rs
