+++ packages/ec-core/src/performance/mod.rs
@@ -5,7 +5,7 @@
 pub mod error_value;
 pub mod score_value;
 
-pub mod summation_behavior;
+pub mod accumulation;
 
 // I don't think we ever _use_ `TestResult` (singular), so maybe it goes away.
 pub mod test_result;
